refactor(invoices): extract share and public access lifecycle
