+++ backend/app/features/invoices/share/__init__.py
@@ -1,0 +1,5 @@
+"""Invoice share lifecycle package."""
+
+from app.features.invoices.share.service import InvoiceShareService
+
+__all__ = ["InvoiceShareService"]
